Allow Desktop OpenGL builds of Qt
Since OpenGL ES2 is just a subset of OpenGL and Qt Quick is not using
features from Desktop OpenGL it is safe to allow non-ES2 versions of Qt.
In case of a requested function is not available a message will be shown
and a null function is returned in getProcAddress function.

macOS cannot be allowed because it is not possible to access the
QCocoaEventDispactcher from another platform plugin.

Change-Id: I55c51c31e2b78978a6dcb6cc4d99ec7502392b40
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>