Allow users to disable the loading/waiting screen animation
Allows the user to disable the browser loading animation for
debugging purposes. WebGL Inspector tries to debug the loading
screen instead of the proper WebGL context.
It can be disabled setting the QT_WEBGL_LOADINGSCREEN environment
variable to '0' or sending the 'noloadingscreen' parameter to the QPA
plugin.
By default, the browser will show the loading screen.

Change-Id: I5d84173d79c3f3fc67b776ba2f8fc8928a0194a1
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>