Generate metadata for function parameters via variadic macros
I changed the QWEBGL_FUNCTION macro, so metadata for function
parameters like name and type are available at runtime.
This data can be sent once to the client, so it does not need
to be retransmitted for each function call.
I also made macro QWEBGL_FUNCTION_POSTEVENT which reduces
redundancies for non-waiting OpenGL commands.

Change-Id: I84a3b151feb9ba8d2e8d211de5f8798d0c4dabf4
Reviewed-by: Jesus Fernandez <Jesus.Fernandez@qt.io>