Allow customizing the maximum texture size
Adds a new environment variable,  QT_WEBGL_MAX_TEXTURE_SIZE, to
configure the maximum texture size. The maximum value was hardcoded to
512 to save bandwidth, but it was giving problems when a high number of
glyphs need to be stored in the GPU.
After this patch, the maximum texture size allowed by WebGL will be
used.
The user can set the value to a custom value to optimize the bandwidth
if the default value is too high, text rendering issues can happen.

[ChangeLog][QWebGLContext] Fix text rendering issues when
a lot of different glyphs are used.
[ChangeLog][QWebGLContext] QT_WEBGL_MAX_TEXTURE_SIZE
environment variable added to customize the maximum texture size.

Task-number: QTBUG-66682
Change-Id: Id8a46daf4718a76fcf7d6a2c2cf831f3d6db372d
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>