Remove old partial binary message workaround
The original implementation of the plugin was created using Qt 5.8. A
workaround was added to avoid partial binary messages.

Apparently, this is not happening anymore, so the workaround is not
needed anymore.

Removing the workaround can provide a performance boost since the
application sends the binary message in a single message and the
browser does not need to rebuild the message.

Change-Id: I18e7a1c559d8a5d677ab66131098a6a1e855d985
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>