Blacklist Linux QEMU
The test is failing because unknown reasons. It seenms the process
spawning takes longer time each new test case until it reaches a
QTRY_* timeout. Right now the test suite is failing in the
launcher.qml test case but it seems it's related to a degradation
after running some tests before this one.

Task-number: QTBUG-70670
Change-Id: I4a17abafc7ff8af52ab80ca0bef909d35f09397c
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>