Pass the timestamp as a string from JS to Qt to preserve precision
When the timestamp was passed as a double from JS to Qt while being
hosted on an embedded Linux device it was losing the precision
and as a result scrolling for a ListView does not work because it
does not see the timestamp has having changed. So by passing it as a
string instead means that it does not lose anything in the transition.

Fixes: QTBUG-79842
Change-Id: Ia6bb8b713c0c5296a046ff3f7fddbc8e8249bbd6
Reviewed-by: Jesus Fernandez <jsfdez@gmail.com>