Don't send mouse events if we have touch support
Unfortunately, using stopPropagation() and preventDefault() does
not stop the mouse events from being sent for a touch. So we have
to assume that they will be handled by Qt one way or the other and
only send mouse events if we can tell that we are not on a touch
screen or that the mouse event was not synthesized by WebGL.

This was tested on a Windows host with an iPad and Microsoft Edge
to verify it works in both cases correctly.

Change-Id: Ia2297062b2e1a4894f52954dd567bab5b7df8fa1
Reviewed-by: Jesus Fernandez <jsfdez@gmail.com>