Update commercial license headers
Updated header.COMM to the files in tqtc-qtwebglplugin. Examples, tests
or documentation folders are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a change
in the lin count causes failures in tests.

Task-number: QTQAINFRA-4222
Change-Id: I7ce1d45a365ab0a675793a26818f73aabd685e19
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>